Add Ralph.IRC.Client
